Refactor git utility names
