+++ clowder/git/printing.py
@@ -8,7 +8,7 @@ import sys
 from termcolor import colored, cprint
 
 import clowder.utility.formatting as fmt
-from clowder.git.git_repo import GitRepo
+from clowder.git.repo import GitRepo
 from clowder.utility.execute import execute_command
